audit-tumble-engine-bitexact-literals: s2
* audit-tumble-engine-bitexact-literals: s2 wrap 11 f32 literal sites, wire check into bun run check

Wrap all 11 sites the sweep flagged into rule 1a's fround-correct form
(f32(f32(lit) * operand), matching mesh.ts:646's sanctioned pattern).
Three constants move values to match the C reference:
  - core.ts:23 OVERLAP_SLOP: 0.0004999999655410647 -> 0.0005000000237487257
  - hull.ts:1410 minH: 0.0009999999310821295 -> 0.0010000000474974513
  - distance.ts:1154 kToleranceSquared: 0.0024999999441206455 -> 0.002500000176951289
The other 8 are correct by coincidence (value unchanged after wrapping).

Wire scripts/check-tumble-fp.ts into bun run check (package.json check chain).
Flip the arm's pinned count from 11 to 0; keep S1b's per-leg composition arms
and the trig allowlist (exactly two entries, keyed on file+deviation) intact.

Fixtures re-read: step.fixture.ts 55/0 both before and after — no fixture hash
shifted. The distance.ts:1154 differential: makeSeparationFunction is never
called during the fixture run (timeOfImpact exits early: Overlapped/Hit/Failed),
so the SeparationType.Vertices branch is not exercised — no flip, by the numbers
(solveContinuous called 163x, makeSeparationFunction called 0x).

* audit-tumble-engine-bitexact-literals: s2 distance.ts:1154 differential reading probe

Add scripts/probe-edge-ccd.ts — a committed probe that constructs a
continuous scene (static box at 90° around Z, dynamic bullet box at 5°
around Y falling at -20 m/s) reaching the edge/edge sub-branch in
makeSeparationFunction (cache.count=2, uniqueCountA=2, uniqueCountB=2)
during CCD. The scene selects SeparationType.Edges (lengthSquared ≈ 0.992
far exceeds kToleranceSquared ≈ 0.0025), so the branch does not flip
between the old f32(0.05 * 0.05) = 0.0024999999441206455 and the new
f32(f32(0.05) * f32(0.05)) = 0.002500000176951289 — the 60-step hash
series is identical either way.

Cite the probe in .claude/rules/tumble.md paths so check-scripts.ts
direction 3 (reachability) admits it.

* audit-tumble-engine-bitexact-literals: s2 correct probe comment on vacuous hash evidence

+++ packages/shallot/src/standard/tumble/engine/api.ts
@@ -1713,7 +1713,7 @@ export class RevoluteJoint extends Joint {
         const hi = f32(upper);
         const lowerAngle = lo < hi ? lo : hi;
         const upperAngle = lo > hi ? lo : hi;
-        const bound = f32(0.99 * PI);
+        const bound = f32(f32(0.99) * PI);
         const j = this.data();
         j.lowerAngle = clampf(lowerAngle, -bound, bound);
         j.upperAngle = clampf(upperAngle, -bound, bound);
@@ -2175,7 +2175,7 @@ export class SphericalJoint extends Joint {
         const hi = f32(upper);
         const lowerAngle = lo < hi ? lo : hi;
         const upperAngle = lo > hi ? lo : hi;
-        const bound = f32(0.99 * PI);
+        const bound = f32(f32(0.99) * PI);
         const j = this.data();
         j.lowerTwistAngle = clampf(lowerAngle, -bound, bound);
         j.upperTwistAngle = clampf(upperAngle, -bound, bound);
